requirements.txt
streamlit
fpdf2
pytz
